@@ -5,7 +5,6 @@ import org.eclipse.rdf4j.model.Model;
 import org.eclipse.rdf4j.model.Resource;
 import org.eclipse.rdf4j.model.Statement;
 import org.eclipse.rdf4j.model.impl.LinkedHashModel;
-import org.eclipse.rdf4j.model.impl.SimpleIRI;
 import org.eclipse.rdf4j.repository.RepositoryConnection;
 import org.eclipse.rdf4j.repository.RepositoryException;
 import org.eclipse.rdf4j.repository.RepositoryResult;
@@ -44,7 +43,7 @@ public class IndicatorPublication extends RdfService {
 			while (statements.hasNext()) {
 				Statement st = statements.next();
 				// Triplets that don't get published
-				String pred = ((SimpleIRI) st.getPredicate()).toString();
+				String pred = RdfUtils.toString(st.getPredicate());
 				
 				if (pred.endsWith("isValidated")
 						|| pred.endsWith("validationState")) {
