@@ -13,8 +13,12 @@ import org.eclipse.rdf4j.model.IRI;
 import org.eclipse.rdf4j.model.Model;
 import org.eclipse.rdf4j.model.Resource;
 import org.eclipse.rdf4j.model.impl.LinkedHashModel;
-import org.eclipse.rdf4j.model.impl.SimpleIRI;
-import org.eclipse.rdf4j.model.vocabulary.*;
+import org.eclipse.rdf4j.model.vocabulary.DC;
+import org.eclipse.rdf4j.model.vocabulary.DCTERMS;
+import org.eclipse.rdf4j.model.vocabulary.RDF;
+import org.eclipse.rdf4j.model.vocabulary.RDFS;
+import org.eclipse.rdf4j.model.vocabulary.SKOS;
+import org.eclipse.rdf4j.model.vocabulary.XSD;
 import org.json.JSONArray;
 import org.json.JSONObject;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -41,7 +45,10 @@ import fr.insee.rmes.utils.DateUtils;
 
 @Component
 public class StructureComponentUtils extends RdfService {
-    private static final String IO_EXCEPTION = "IOException";
+    private static final String MAX_LENGTH = "maxLength";
+	private static final String MIN_LENGTH = "minLength";
+	private static final String PATTERN = "pattern";
+	private static final String IO_EXCEPTION = "IOException";
 	static final Logger logger = LogManager.getLogger(StructureComponentUtils.class);
     public static final String VALIDATED = "Validated";
     public static final String MODIFIED = "Modified";
@@ -64,7 +71,7 @@ public class StructureComponentUtils extends RdfService {
 
     private void addCodeListRange(JSONObject response) {
         if (response.has(Constants.CODELIST)) {
-            response.put("range", ((SimpleIRI)INSEE.CODELIST).toString());
+            response.put("range", RdfUtils.toString(INSEE.CODELIST));
         }
     }
 
@@ -144,9 +151,9 @@ public class StructureComponentUtils extends RdfService {
         }
 
         String type = component.getType();
-        if (type.equals(((SimpleIRI)QB.ATTRIBUTE_PROPERTY).toString())) {
+        if (type.equals(RdfUtils.toString(QB.ATTRIBUTE_PROPERTY))) {
             createRDFForComponent(component, QB.ATTRIBUTE_PROPERTY, RdfUtils.structureComponentAttributeIRI(component.getId()), status);
-        } else if (type.equals(((SimpleIRI)QB.MEASURE_PROPERTY).toString())) {
+        } else if (type.equals(RdfUtils.toString(QB.MEASURE_PROPERTY))) {
             createRDFForComponent(component, QB.MEASURE_PROPERTY, RdfUtils.structureComponentMeasureIRI(component.getId()), status);
         } else {
             createRDFForComponent(component, QB.DIMENSION_PROPERTY, RdfUtils.structureComponentDimensionIRI(component.getId()), status);
@@ -177,7 +184,7 @@ public class StructureComponentUtils extends RdfService {
             RdfUtils.addTripleUri(componentURI, QB.CONCEPT, INSEE.STRUCTURE_CONCEPT + component.getConcept(), model, graph);
         }
 
-        if (component.getRange() != null && component.getRange().equals(((SimpleIRI)INSEE.CODELIST).toString())) {
+        if (component.getRange() != null && component.getRange().equals(RdfUtils.toString(INSEE.CODELIST))) {
             RdfUtils.addTripleUri(componentURI, RDF.TYPE, QB.CODED_PROPERTY, model, graph);
 
             JSONObject object = repoGestion.getResponseAsObject(StructureQueries.getUriClasseOwl(component.getCodeList()));
@@ -187,31 +194,25 @@ public class StructureComponentUtils extends RdfService {
             } else {
                 RdfUtils.addTripleUri(componentURI, RDFS.RANGE, SKOS.CONCEPT, model, graph);
             }
-        } else {
+        } else if (component.getRange() != null) {
             RdfUtils.addTripleUri(componentURI, RDFS.RANGE, component.getRange(), model, graph);
 
-            if (component.getRange() != null && component.getRange().equals(XSD.DATE.toString())) {
-                RdfUtils.addTripleString(componentURI, RdfUtils.createXSDIRI("pattern"), component.getPattern(), Config.LG1, model, graph);
+            if (component.getRange().equals(XSD.DATE.stringValue())) {
+                RdfUtils.addTripleString(componentURI, RdfUtils.createXSDIRI(PATTERN), component.getPattern(), Config.LG1, model, graph);
             }
-            if (component.getRange() != null && component.getRange().equals(XSD.DATETIME.toString())) {
-                RdfUtils.addTripleString(componentURI, RdfUtils.createXSDIRI("pattern"), component.getPattern(), Config.LG1, model, graph);
+            else if (component.getRange().equals(XSD.DATETIME.stringValue())) {
+                RdfUtils.addTripleString(componentURI, RdfUtils.createXSDIRI(PATTERN), component.getPattern(), Config.LG1, model, graph);
             }
-            if (component.getRange() != null && component.getRange().equals(XSD.INTEGER.toString())) {
-                RdfUtils.addTripleString(componentURI, RdfUtils.createXSDIRI("minLength"), component.getMinLength(), Config.LG1, model, graph);
-                RdfUtils.addTripleString(componentURI, RdfUtils.createXSDIRI("maxLength"), component.getMaxLength(), Config.LG1, model, graph);
+            else if (component.getRange().equals(XSD.INTEGER.stringValue()) || component.getRange().equals(XSD.DOUBLE.stringValue())) {
+                RdfUtils.addTripleString(componentURI, RdfUtils.createXSDIRI(MIN_LENGTH), component.getMinLength(), Config.LG1, model, graph);
+                RdfUtils.addTripleString(componentURI, RdfUtils.createXSDIRI(MAX_LENGTH), component.getMaxLength(), Config.LG1, model, graph);
                 RdfUtils.addTripleString(componentURI, RdfUtils.createXSDIRI("minInclusive"), component.getMinInclusive(), Config.LG1, model, graph);
                 RdfUtils.addTripleString(componentURI, RdfUtils.createXSDIRI("maxInclusive"), component.getMaxInclusive(), Config.LG1, model, graph);
             }
-            if (component.getRange() != null && component.getRange().equals(XSD.DOUBLE.toString())) {
-                RdfUtils.addTripleString(componentURI, RdfUtils.createXSDIRI("minLength"), component.getMinLength(), Config.LG1, model, graph);
-                RdfUtils.addTripleString(componentURI, RdfUtils.createXSDIRI("maxLength"), component.getMaxLength(), Config.LG1, model, graph);
-                RdfUtils.addTripleString(componentURI, RdfUtils.createXSDIRI("minInclusive"), component.getMinInclusive(), Config.LG1, model, graph);
-                RdfUtils.addTripleString(componentURI, RdfUtils.createXSDIRI("maxInclusive"), component.getMaxInclusive(), Config.LG1, model, graph);
-            }
-            if (component.getRange() != null && component.getRange().equals(XSD.STRING.toString())) {
-                RdfUtils.addTripleString(componentURI, RdfUtils.createXSDIRI("minLength"), component.getMinLength(), Config.LG1, model, graph);
-                RdfUtils.addTripleString(componentURI, RdfUtils.createXSDIRI("maxLength"), component.getMaxLength(), Config.LG1, model, graph);
-                RdfUtils.addTripleString(componentURI, RdfUtils.createXSDIRI("pattern"), component.getPattern(), Config.LG1, model, graph);
+            else if (component.getRange().equals(XSD.STRING.stringValue())) {
+                RdfUtils.addTripleString(componentURI, RdfUtils.createXSDIRI(MIN_LENGTH), component.getMinLength(), Config.LG1, model, graph);
+                RdfUtils.addTripleString(componentURI, RdfUtils.createXSDIRI(MAX_LENGTH), component.getMaxLength(), Config.LG1, model, graph);
+                RdfUtils.addTripleString(componentURI, RdfUtils.createXSDIRI(PATTERN), component.getPattern(), Config.LG1, model, graph);
             }
         }
 
@@ -223,10 +224,10 @@ public class StructureComponentUtils extends RdfService {
     }
 
     private String generateNextId(String type) throws RmesException {
-        if (type.equals(((SimpleIRI)QB.ATTRIBUTE_PROPERTY).toString())) {
+        if (type.equals(RdfUtils.toString(QB.ATTRIBUTE_PROPERTY))) {
             return generateNextId("a", "attribut", QB.ATTRIBUTE_PROPERTY);
         }
-        if (type.equals(((SimpleIRI)QB.MEASURE_PROPERTY).toString())) {
+        if (type.equals(RdfUtils.toString(QB.MEASURE_PROPERTY))) {
             return generateNextId("m", "mesure", QB.MEASURE_PROPERTY);
         }
         return generateNextId("d", "dimension", QB.DIMENSION_PROPERTY);
@@ -236,7 +237,7 @@ public class StructureComponentUtils extends RdfService {
 
     private String generateNextId(String prefix, String namespaceSuffix, IRI type) throws RmesException {
         logger.info("Generate id for component");
-        JSONObject json = repoGestion.getResponseAsObject(StructureQueries.lastId(namespaceSuffix, ((SimpleIRI)type).toString()));
+        JSONObject json = repoGestion.getResponseAsObject(StructureQueries.lastId(namespaceSuffix, RdfUtils.toString(type)));
         logger.debug("JSON when generating the id of a component : {}", json);
         if (json.length() == 0) {
             return prefix + "1000";
@@ -298,9 +299,9 @@ public class StructureComponentUtils extends RdfService {
             throw new RmesException(ErrorCodes.COMPONENT_FORBIDDEN_DELETE, "You cannot delete a validated component", new JSONArray());
         }
         IRI componentIri;
-        if (type.equalsIgnoreCase(((SimpleIRI)QB.ATTRIBUTE_PROPERTY).toString())) {
+        if (type.equalsIgnoreCase(QB.ATTRIBUTE_PROPERTY.stringValue())) {
             componentIri =  RdfUtils.structureComponentAttributeIRI(id);
-        } else if (type.equalsIgnoreCase(((SimpleIRI)QB.MEASURE_PROPERTY).toString())) {
+        } else if (type.equalsIgnoreCase(QB.MEASURE_PROPERTY.stringValue())) {
             componentIri =  RdfUtils.structureComponentMeasureIRI(id);
         } else {
             componentIri =  RdfUtils.structureComponentDimensionIRI(id);
@@ -338,13 +339,13 @@ public class StructureComponentUtils extends RdfService {
         String type = component.getString("type");
         String id = component.getString("id");
 
-        if (type.equals(((SimpleIRI)QB.ATTRIBUTE_PROPERTY).toString())) {
+        if (type.equals(QB.ATTRIBUTE_PROPERTY.stringValue())) {
             componentPublication.publishComponent(RdfUtils.structureComponentAttributeIRI(id), QB.ATTRIBUTE_PROPERTY);
         }
-        else if (type.equals(((SimpleIRI)QB.MEASURE_PROPERTY).toString())) {
+        else if (type.equals(QB.MEASURE_PROPERTY.stringValue())) {
             componentPublication.publishComponent(RdfUtils.structureComponentMeasureIRI(id), QB.MEASURE_PROPERTY);
         }
-        else if (type.equals(((SimpleIRI)QB.DIMENSION_PROPERTY).toString())) {
+        else if (type.equals(QB.DIMENSION_PROPERTY.stringValue())) {
             componentPublication.publishComponent(RdfUtils.structureComponentDimensionIRI(id), QB.DIMENSION_PROPERTY);
         }
 
