@@ -45,12 +45,12 @@ public class ApplicationContext {
 		return HttpClients.custom().setSSLHostnameVerifier(NoopHostnameVerifier.INSTANCE).setSSLSocketFactory(sslsf);
 	}
 	
-	@Bean
-	public RestTemplate restTemplate() throws KeyManagementException, KeyStoreException, NoSuchAlgorithmException {
-		CloseableHttpClient httpClient = httpClientBuilder().build();
-		ClientHttpRequestFactory requestFactory = new HttpComponentsClientHttpRequestFactory(httpClient);
-		return new RestTemplate(requestFactory);
-	}
+//	@Bean
+//	public RestTemplate restTemplate() throws KeyManagementException, KeyStoreException, NoSuchAlgorithmException {
+//		CloseableHttpClient httpClient = httpClientBuilder().build();
+//		ClientHttpRequestFactory requestFactory = new HttpComponentsClientHttpRequestFactory(httpClient);
+//		return new RestTemplate(requestFactory);
+//	}
 
 	@PostConstruct
 	public void setUp() {
