@@ -1,8 +1,8 @@
 package fr.insee.rmes.external_services.authentication.user_roles_manager;
 
-import java.io.StringReader;
 import java.text.MessageFormat;
-import java.util.List;
+import java.util.HashMap;
+import java.util.Map;
 import java.util.TreeSet;
 
 import javax.naming.NamingEnumeration;
@@ -18,43 +18,52 @@ import org.apache.commons.lang3.StringUtils;
 import org.apache.http.HttpStatus;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
-import org.jdom.Document;
-import org.jdom.Element;
-import org.jdom.input.SAXBuilder;
-import org.jdom.xpath.XPath;
+import org.glassfish.jersey.client.authentication.HttpAuthenticationFeature;
 import org.json.JSONArray;
 import org.json.JSONObject;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+
+import com.fasterxml.jackson.core.JsonProcessingException;
+import com.fasterxml.jackson.databind.ObjectMapper;
 
 import fr.insee.rmes.bauhaus_services.Constants;
 import fr.insee.rmes.config.Config;
 import fr.insee.rmes.config.auth.roles.UserRolesManagerService;
 import fr.insee.rmes.exceptions.RmesException;
 import fr.insee.rmes.external_services.authentication.LdapConnexion;
+import fr.insee.rmes.external_services.authentication.user_roles_manager.sugoiModel.Application;
+import fr.insee.rmes.external_services.authentication.user_roles_manager.sugoiModel.Group;
+import fr.insee.rmes.external_services.authentication.user_roles_manager.sugoiModel.UserSugoi;
+import fr.insee.rmes.external_services.authentication.user_roles_manager.sugoiModel.UsersSugoi;
 import fr.insee.rmes.utils.JSONComparator;
 
 @Service
 public class RmesUserRolesManagerImpl implements UserRolesManagerService {
 	
+	private static final int NB_USERS_EXPECTED = 20000;
+
 	@Autowired
 	LdapConnexion ldapConnexion;
 
 	static final Logger logger = LogManager.getLogger(RmesUserRolesManagerImpl.class);
 
 	private static final String IGESA_APP_SEARCH_PATH = "/recherche/application/";
+	private static final String SUGOI_REALM_SEARCH_PATH = "/realms/";
+	private static final String SUGOI_APP_SEARCH_PATH = "/applications/";
+	private static final String SUGOI_SEARCH_APP = Config.SUGOI_URL + SUGOI_REALM_SEARCH_PATH + Config.SUGOI_REALM + SUGOI_APP_SEARCH_PATH + Config.SUGOI_APP ;
+	private static final String SUGOI_SEARCH_USERS = Config.SUGOI_URL + SUGOI_REALM_SEARCH_PATH + Config.SUGOI_REALM + "/users" ;
+
 
 	private static final String IGESA_ADD_USER_PATH_FMT = Config.IGESA_URL + "/gestion/ajout/personne/application/"
 			+ Config.IGESA_APP_ID + "/groupe/{1}/utilisateur/{0}";
 	private static final String IGESA_DELETE_USER_PATH_FMT = Config.IGESA_URL
 			+ "/gestion/suppression/personne/application/" + Config.IGESA_APP_ID + "/groupe/{1}/utilisateur/{0}";
 
-	private static final String ROLES_XPATH = "/applications/application/groupes/groupe";
 	private static final String ROLE_ID_XPATH = "cn";
-	private static final String ROLE_LABEL_XPATH = "description";
-	private static final String ROLE_PERSONS_XPATH = "personnes/personne";
-	private static final String ROLE_PERSON_STAMP_XPATH = "ou";
 	private static final String ROLE_PERSON_IDEP_XPATH = "uid";
+	
+	private Map<String,UserSugoi> mapUsers;
 	
 	@Override
 	public String getAuth(String body) {
@@ -68,33 +77,37 @@ public class RmesUserRolesManagerImpl implements UserRolesManagerService {
 	}
 
 	@Override
-	@SuppressWarnings("unchecked")
 	public String getRoles() throws RmesException {
+		if (mapUsers == null || mapUsers.isEmpty()) {getAgentsSugoi();}
+		logger.info("mapUsers size : {} / {} max", mapUsers.size(), NB_USERS_EXPECTED);
 		JSONArray roles = new JSONArray();
 		try {
-			Client client = ClientBuilder.newClient();
+			Client client = ClientBuilder.newClient().register(HttpAuthenticationFeature.basic(Config.SUGOI_USER, Config.SUGOI_PASSWORD));	
+			String jsonResponse = client.target(SUGOI_SEARCH_APP).request(MediaType.APPLICATION_JSON).get(String.class);
 
-			String xmlResponse = client.target(Config.IGESA_URL + IGESA_APP_SEARCH_PATH + Config.IGESA_APP_ID)
-					.request(MediaType.APPLICATION_XML).get(String.class);
-
-			Document doc = new SAXBuilder().build(new StringReader(xmlResponse));
-			List<Element> l = (XPath.selectNodes(doc, ROLES_XPATH));
-			for (Element e : l) {
-				JSONObject jsonO = new JSONObject();
-				jsonO.put(Constants.ID, XPath.newInstance(ROLE_ID_XPATH).valueOf(e));
-				jsonO.put(Constants.LABEL, XPath.newInstance(ROLE_LABEL_XPATH).valueOf(e));
-				List<Element> p = (XPath.selectNodes(e, ROLE_PERSONS_XPATH));
+			ObjectMapper mapper = new ObjectMapper();
+			Application application = mapper.readValue(jsonResponse, Application.class);
+			for (Group g : application.getGroups()) {
+				JSONObject jsonGroup = new JSONObject();
+				jsonGroup.put(Constants.ID, g.getName());
+				jsonGroup.put(Constants.LABEL, g.getDescription());
 				JSONArray persons = new JSONArray();
-				for (Element person : p) {
-					JSONObject jsonOO = new JSONObject();
-					jsonOO.put(Constants.ID, XPath.newInstance(ROLE_PERSON_IDEP_XPATH).valueOf(person));
-					jsonOO.put(Constants.LABEL, XPath.newInstance(ROLE_ID_XPATH).valueOf(person));
-					jsonOO.put("stamp", XPath.newInstance(ROLE_PERSON_STAMP_XPATH).valueOf(person));
-					persons.put(jsonOO);
+				if (g.getUsers() != null) {
+					for (UserSugoi u : g.getUsers()) {
+						UserSugoi completeUser = mapUsers.get(u.getUsername().toLowerCase());
+						if (completeUser != null ) {
+							JSONObject jsonUser = new JSONObject();
+							jsonUser.put(Constants.ID, u.getUsername());
+							jsonUser.put(Constants.LABEL, completeUser.getCompleteName());
+							jsonUser.put(Constants.STAMP, completeUser.getAttributes().getInseeTimbre());
+							persons.put(jsonUser);
+						}	else logger.warn("Unknown user : {}",u.getUsername());
+					}
 				}
-				jsonO.put("persons", persons);
-				roles.put(jsonO);
+				jsonGroup.put("persons", persons);
+				roles.put(jsonGroup);
 			}
+			
 		} catch (Exception e) {
 			throw new RmesException(HttpStatus.SC_INTERNAL_SERVER_ERROR, e.getMessage(), "Fail to getRoles");
 		}
@@ -132,6 +145,38 @@ public class RmesUserRolesManagerImpl implements UserRolesManagerService {
 		}
 		return agents.toString();
 	}
+	
+	
+	public String getAgentsSugoi() throws RmesException {
+		mapUsers = new HashMap<>(NB_USERS_EXPECTED);
+		TreeSet<JSONObject> agents = new TreeSet<>(new JSONComparator(Constants.LABEL));
+
+		Client client = ClientBuilder.newClient().register(HttpAuthenticationFeature.basic(Config.SUGOI_USER, Config.SUGOI_PASSWORD));	
+		String jsonResponse = client.target(SUGOI_SEARCH_USERS)
+									.queryParam("size", NB_USERS_EXPECTED)
+									.request(MediaType.APPLICATION_JSON)
+									.get(String.class);
+
+		ObjectMapper mapper = new ObjectMapper();
+		UsersSugoi users;
+		try {
+			users = mapper.readValue(jsonResponse, UsersSugoi.class);
+			for (UserSugoi u : users.getResults()) {
+				JSONObject jsonUser = new JSONObject();
+				jsonUser.put(Constants.ID, u.getUsername());
+				jsonUser.put(Constants.LABEL, u.getCompleteName());
+				if (u.getAttributes() != null) {
+					jsonUser.put(Constants.STAMP, u.getAttributes().getInseeTimbre());
+				}
+				agents.add(jsonUser);
+				mapUsers.put(u.getUsername().toLowerCase(), u);
+			}
+		} catch (JsonProcessingException e) {
+			logger.error("Get agents via Sugoi failed : {}", e.getMessage());
+			throw new RmesException(HttpStatus.SC_INTERNAL_SERVER_ERROR, e.getMessage(), "Get agents via Sugoi failed");
+		}
+		return agents.toString();
+	}
 
 	@Override
 	public void setAddRole(String role, String user) {
@@ -158,6 +203,20 @@ public class RmesUserRolesManagerImpl implements UserRolesManagerService {
 			throw new RmesException(HttpStatus.SC_INTERNAL_SERVER_ERROR, e.getMessage(), "Fail to target IGESA");
 		}
 		return StringUtils.isEmpty(xmlResponse)? "KO" :  "OK";
+	}
+	
+	@Override
+	public String checkSugoiConnexion() throws RmesException {
+		String jsonResponse ="";
+		try {
+			Client client = ClientBuilder.newClient().register(HttpAuthenticationFeature.basic(Config.SUGOI_USER, Config.SUGOI_PASSWORD));
+
+			jsonResponse = client.target(Config.SUGOI_URL + "whoami")
+					.request(MediaType.APPLICATION_JSON).get(String.class);
+		} catch (Exception e) {
+			throw new RmesException(HttpStatus.SC_INTERNAL_SERVER_ERROR, e.getMessage(), "Fail to target SUGOI");
+		}
+		return StringUtils.isEmpty(jsonResponse)? "KO" :  "OK";
 	}
 
 }
