@@ -1,30 +1,33 @@
 package fr.insee.rmes.persistance.sparql_queries.code_list;
 
-import fr.insee.rmes.bauhaus_services.ConceptsService;
-import fr.insee.rmes.bauhaus_services.Constants;
+import java.util.HashMap;
+
 import fr.insee.rmes.bauhaus_services.rdf_utils.FreeMarkerUtils;
 import fr.insee.rmes.config.Config;
 import fr.insee.rmes.exceptions.RmesException;
 
-import java.util.HashMap;
-
 public class CodeListQueries {
+
+	private static final String CODES_LISTS_GRAPH = "CODES_LISTS_GRAPH";
+	private static final String CODES_LIST = "codes-list/";
+	private static final String PARTIAL = "PARTIAL";
+	private static final String NOTATION = "NOTATION";
 
 	public static String isCodesListValidated(String codesListUri) throws RmesException {
 		HashMap<String, Object> params = new HashMap<>();
-		params.put("CODES_LISTS_GRAPH", Config.CODELIST_GRAPH);
+		params.put(CODES_LISTS_GRAPH, Config.CODELIST_GRAPH);
 		params.put("IRI", codesListUri);
-		return FreeMarkerUtils.buildRequest("codes-list/", "isCodesListValidated.ftlh", params);
+		return FreeMarkerUtils.buildRequest(CODES_LIST, "isCodesListValidated.ftlh", params);
 	}
 
 	public static String getAllCodesLists(boolean partial) throws RmesException {
 
 		HashMap<String, Object> params = new HashMap<>();
-		params.put("CODES_LISTS_GRAPH", Config.CODELIST_GRAPH);
+		params.put(CODES_LISTS_GRAPH, Config.CODELIST_GRAPH);
 		params.put("LG1", Config.LG1);
 		params.put("LG2", Config.LG2);
-		params.put("PARTIAL", partial);
-		return FreeMarkerUtils.buildRequest("codes-list/", "getAllCodesLists.ftlh", params);
+		params.put(PARTIAL, partial);
+		return FreeMarkerUtils.buildRequest(CODES_LIST, "getAllCodesLists.ftlh", params);
 	}
 	public static String getCodeListItemsByNotation(String notation) {
 		return "SELECT ?code ?labelLg1 ?labelLg2 \n"
@@ -78,48 +81,48 @@ public class CodeListQueries {
 
 	public static String geCodesListByIRI(String id) throws RmesException {
 		HashMap<String, Object> params = new HashMap<>();
-		params.put("CODES_LISTS_GRAPH", Config.CODELIST_GRAPH);
+		params.put(CODES_LISTS_GRAPH, Config.CODELIST_GRAPH);
 		params.put("CODE_LIST", id);
 		params.put("LG1", Config.LG1);
 		params.put("LG2", Config.LG2);
-		return FreeMarkerUtils.buildRequest("codes-list/", "getCodeListByIRI.ftlh", params);
+		return FreeMarkerUtils.buildRequest(CODES_LIST, "getCodeListByIRI.ftlh", params);
 	}
 
 	public static String getDetailedCodeListByNotation(String notation) throws RmesException {
 		HashMap<String, Object> params = getInitParams();
-		params.put("NOTATION", notation);
+		params.put(NOTATION, notation);
 		params.put("CODE_LIST_BASE_URI", Config.CODE_LIST_BASE_URI);
-		return FreeMarkerUtils.buildRequest("codes-list/", "getDetailedCodesList.ftlh", params);
+		return FreeMarkerUtils.buildRequest(CODES_LIST, "getDetailedCodesList.ftlh", params);
 	}
 
 	public static String getCodesListsForSearch(boolean partial) throws RmesException {
 		HashMap<String, Object> params = getInitParams();
-		params.put("PARTIAL", partial);
-		return FreeMarkerUtils.buildRequest("codes-list/", "getDetailedCodesListForSearch.ftlh", params);
+		params.put(PARTIAL, partial);
+		return FreeMarkerUtils.buildRequest(CODES_LIST, "getDetailedCodesListForSearch.ftlh", params);
 	}
 
 	public static String getCodesForSearch(boolean partial) throws RmesException {
 		HashMap<String, Object> params = getInitParams();
-		params.put("PARTIAL", partial);
-		return FreeMarkerUtils.buildRequest("codes-list/", "getCodesForSearch.ftlh", params);
+		params.put(PARTIAL, partial);
+		return FreeMarkerUtils.buildRequest(CODES_LIST, "getCodesForSearch.ftlh", params);
 	}
 
 	public static String getDetailedCodes(String notation, boolean partial) throws RmesException {
 		HashMap<String, Object> params = getInitParams();
-		params.put("NOTATION", notation);
-		params.put("PARTIAL", partial);
-		return FreeMarkerUtils.buildRequest("codes-list/", "getDetailedCodes.ftlh", params);
+		params.put(NOTATION, notation);
+		params.put(PARTIAL, partial);
+		return FreeMarkerUtils.buildRequest(CODES_LIST, "getDetailedCodes.ftlh", params);
 	}
 
 	public static String getCodesSeq(String notation) throws RmesException {
 		HashMap<String, Object> params = getInitParams();
-		params.put("NOTATION", notation);
-		return FreeMarkerUtils.buildRequest("codes-list/", "getCodesSeq.ftlh", params);
+		params.put(NOTATION, notation);
+		return FreeMarkerUtils.buildRequest(CODES_LIST, "getCodesSeq.ftlh", params);
 	}
 
 	private static HashMap<String, Object> getInitParams() {
 		HashMap<String, Object> params = new HashMap<>();
-		params.put("CODES_LISTS_GRAPH", Config.CODELIST_GRAPH);
+		params.put(CODES_LISTS_GRAPH, Config.CODELIST_GRAPH);
 		params.put("LG1", Config.LG1);
 		params.put("LG2", Config.LG2);
 		return params;
@@ -135,8 +138,8 @@ public class CodeListQueries {
 		params.put("ID", id);
 		params.put("IRI", iri);
 		params.put("SEE_ALSO", seeAlso);
-		params.put("PARTIAL", partial);
+		params.put(PARTIAL, partial);
 
-		return FreeMarkerUtils.buildRequest("codes-list/", "checkCodeListUnicity.ftlh", params);
+		return FreeMarkerUtils.buildRequest(CODES_LIST, "checkCodeListUnicity.ftlh", params);
 	}
 }
